Adding integration suite (#6)

+++ models/src/main/scala/com/github/nullptr7/models.scala
@@ -3,11 +3,10 @@ package com.github.nullptr7
 import io.circe.Codec
 import io.circe.generic.semiauto.deriveCodec
 
-
 package models {
 
   final case class Employee(
-    id:      Long,
+    id:      Int,
     name:    String,
     age:     Int,
     salary:  Double,
@@ -15,7 +14,7 @@ package models {
   )
 
   final case class Address(
-    id:     Long,
+    id:     Int,
     street: String,
     city:   String,
     state:  String,
